+++ lib/src/sqlite/DBInterface.dart
@@ -23,11 +23,22 @@
 ///
 import 'dart:async' show Future;
 
-import 'package:sqflite/sqflite.dart' show Database, DatabaseException, OnDatabaseConfigureFn, OnDatabaseCreateFn, OnDatabaseOpenFn, OnDatabaseVersionChangeFn, openDatabase;
+import 'dart:io' show Directory;
+
+import 'package:sqflite/sqflite.dart'
+    show
+        Database,
+        DatabaseException,
+        OnDatabaseConfigureFn,
+        OnDatabaseCreateFn,
+        OnDatabaseOpenFn,
+        OnDatabaseVersionChangeFn,
+        openDatabase;
 
 import 'package:path/path.dart' show join;
 
-import 'package:path_provider/path_provider.dart' show getApplicationDocumentsDirectory;
+import 'package:path_provider/path_provider.dart'
+    show getApplicationDocumentsDirectory;
 
 import 'package:flutter/foundation.dart' show mustCallSuper;
 
@@ -99,21 +110,23 @@ abstract class DBInterface {
     return open;
   }
 
-  close() => _dbInt.close();
+  close() {
+    _dbInt.close();
+  }
 
   /// List of the tables and list their fields: Map<String, List>
   Map<String, List> get fields => _dbInt._fields;
 
-  /// List of the table and a map of each of their field values: Map<String, Map>
-  Map<String, Map> get values => _dbInt._fldValues;
+  /// Get the key field for a table
+  Future<String> keyField(String table) async {
+    if (db == null) await open();
+    return _dbInt._keyFields[table];
+  }
 
   Map<String, Map> get newrec => _dbInt._newRec;
 
   /// Gets the Database
   Database get db => _dbInt.db;
-
-  /// Key field
-  String get keyField => _dbInt.keyField;
 
   /// Gets the exception if any.
   Exception get error => _dbError.e;
@@ -146,27 +159,17 @@ abstract class DBInterface {
   /// How many records were last updated.
   int get recsUpdated => _dbInt.rowsUpdated;
 
-  Map addEntries(String table, Map values){
-    if(values == null) return {};
-    Map fldValues = _dbInt._fldValues[table];
-    if(fldValues == null) return {};
-    for(String key in values.keys){
-      if(!fldValues.containsKey(key)) continue;
-      fldValues[key] = values[key];
-    }
-    return fldValues;
+  Future<Map<String, dynamic>> saveRec(
+      String table, Map<String, dynamic> fldValues) async {
+    return updateRec(table, fldValues);
   }
 
-  Future<Map> saveRec(String table) async {
-    return updateRec(table, _dbInt._fldValues[table]);
-  }
-
-  Future<bool> saveMap(String table, Map values) async{
-    if(table == null || table.isEmpty) Future.value(false);
-    Map rec = newRec(table, values);
-    addEntries(table, rec);
-    rec = await saveRec(table);
-    return rec.isNotEmpty;
+  Future<Map<String, dynamic>> saveMap(
+      String table, Map<String, dynamic> values) async {
+    if (table == null || table.isEmpty) Future.value(Map<String, dynamic>());
+    Map<String, dynamic> rec = newRec(table, values);
+    rec = await saveRec(table, rec);
+    return rec;
   }
 
   Future<bool> runTxn(Func func) async {
@@ -178,25 +181,29 @@ abstract class DBInterface {
     return ret;
   }
 
-  Future<Map> updateRec(String table, Map fields) async {
-    Map rec;
+  Future<Map<String, dynamic>> updateRec(
+      String table, Map<String, dynamic> fields) async {
+    Map<String, dynamic> rec;
     try {
       rec = await _dbInt.updateRec(table, fields);
       _dbError.clear();
     } catch (e) {
       _dbError.set(e);
-      rec = Map();
+      rec = Map<String, dynamic>();
     }
     return rec;
   }
 
   /// Return an 'empty' record map
-  Map newRec(String table, [Map data]) {
-    Map newRec = Map();
+  Map<String, dynamic> newRec(String table, [Map<String, dynamic> data]) {
+    Map<String, dynamic> newRec = Map();
 
     newRec.addAll(_dbInt._newRec[table]);
 
-    if (data != null) newRec.addEntries(data.entries);
+    if (data != null)
+      data.forEach((key, value) {
+        if (newRec.containsKey(key)) newRec[key] = value;
+      });
 
     return newRec;
   }
@@ -301,9 +308,9 @@ abstract class DBInterface {
     return recs;
   }
 
-  List<Map<String, dynamic>> mapQuery(List<Map<String, dynamic>> query){
+  List<Map<String, dynamic>> mapQuery(List<Map<String, dynamic>> query) {
     List<Map<String, dynamic>> mapList = [];
-    for(var row in query){
+    for (var row in query) {
       Map<String, dynamic> map = row.map((key, value) => MapEntry(key, value));
       mapList.add(map);
     }
@@ -337,7 +344,7 @@ abstract class DBInterface {
   static Exception _exception;
 
   static setError(Object ex) {
-    if(ex is! Exception) ex = Exception(ex.toString());
+    if (ex is! Exception) ex = Exception(ex.toString());
     _exception = ex;
   }
 
@@ -434,11 +441,7 @@ class _DBInterface {
   final OnDatabaseVersionChangeFn onUpgrade;
   final OnDatabaseVersionChangeFn onDowngrade;
 
-  // ROWID is automatically added to all SQLite tables by default, and is a unique integer,
-  String keyField = 'rowid';
-
   Exception ex;
-
   Database db;
 
   Future<bool> open() async {
@@ -465,7 +468,7 @@ class _DBInterface {
         );
 
         // Create the Map objects containing the table's fields.
-        tableFields();
+        await tableFields();
 
         opened = true;
       } catch (e) {
@@ -478,33 +481,41 @@ class _DBInterface {
 
   void close() {
     var temp = db;
+
     /// Hot reload must have db close & set to null first.
     db = null;
+
     /// Don't provide the 'await' command as the process MUST wait.
     temp?.close();
   }
 
   int rowsUpdated;
 
-  Future<Map> updateRec(String table, Map fields) async {
+  Future<Map<String, dynamic>> updateRec(
+      String table, Map<String, dynamic> fields) async {
     rowsUpdated = 0;
-    if (fields[keyField] == null) {
-      fields[keyField] = await db.insert(table, fields);
+    String keyFld = _keyFields[table];
+    var keyValue = fields[keyFld];
+    if (table == null) {
+      /// We got nothing.
+    } else if (keyValue == null) {
+      fields[keyFld] = await db.insert(table, fields);
       rowsUpdated = 1;
     } else {
-      rowsUpdated = await db.update(table, fields,
-          where: "$keyField = ?", whereArgs: [fields[keyField]]);
+      rowsUpdated = await db
+          .update(table, fields, where: "$keyFld = ?", whereArgs: [keyValue]);
     }
     return fields;
   }
 
-  Future<List<Map>> getRec(String table, int id, List fields) async {
+  Future<List<Map<String, dynamic>>> getRec(
+      String table, int id, List fields) async {
     if (db == null) {
       final open = await this.open();
       if (!open) return Future.value([{}]);
     }
-    return await db
-        .query(table, columns: fields, where: "$keyField = ?", whereArgs: [id]);
+    return await db.query(table,
+        columns: fields, where: "${_keyFields[table]} = ?", whereArgs: [id]);
   }
 
   Future<int> delete(String table, int id) async {
@@ -512,7 +523,8 @@ class _DBInterface {
       final open = await this.open();
       if (!open) return Future.value(0);
     }
-    return await db.delete(table, where: "$keyField = ?", whereArgs: [id]);
+    return await db
+        .delete(table, where: "${_keyFields[table]} = ?", whereArgs: [id]);
   }
 
   Future<List<Map<String, dynamic>>> rawQuery(String sqlStmt) async {
@@ -523,7 +535,7 @@ class _DBInterface {
     return await db.rawQuery(sqlStmt);
   }
 
-  Future<List<Map>> query(String table,
+  Future<List<Map<String, dynamic>>> query(String table,
       {bool distinct = false,
       List<String> columns,
       String where,
@@ -564,7 +576,7 @@ class _DBInterface {
     }
   }
 
-  Future<List<Map>> tableNames() async {
+  Future<List<Map<String, dynamic>>> tableNames() async {
     if (db == null) {
       final open = await this.open();
       if (!open) return Future.value([{}]);
@@ -573,7 +585,7 @@ class _DBInterface {
         .rawQuery("SELECT name FROM sqlite_master WHERE type='table'");
   }
 
-  Future<List<Map>> tableColumns(String table) async {
+  Future<List<Map<String, dynamic>>> tableColumns(String table) async {
     if (db == null) {
       final open = await this.open();
       if (!open) return Future.value([{}]);
@@ -582,7 +594,7 @@ class _DBInterface {
   }
 
   Future<List<String>> tableList() async {
-    List<Map> tables = await tableNames();
+    List<Map<String, dynamic>> tables = await tableNames();
 
     List<String> list = List();
 
@@ -594,11 +606,14 @@ class _DBInterface {
 
   final Map<String, List> _fields = Map();
 
-  final Map<String, Map> _fldValues = Map();
+  final Map<String, String> _keyFields = Map();
 
-  final Map<String, Map> _newRec = Map();
+  final Map<String, Map<String, dynamic>> _newRec = Map();
 
-  void tableFields() async {
+  Future<void> tableFields() async {
+    dynamic fldValue;
+    String keyField;
+
     var tables = await tableList();
 
     for (var table in tables) {
@@ -606,17 +621,23 @@ class _DBInterface {
 
       List<String> fields = List();
 
+      /// ROWID is automatically added to all SQLite tables by default, and is a unique integer,
+      keyField = 'rowid';
+
       fields.add(keyField);
 
       Map<String, dynamic> fieldValues = Map();
 
       fieldValues[keyField] = null;
 
+      _keyFields[table] = keyField;
+
       for (var col in columns) {
         /// Replace the primary key field.
         if (col['pk'] == 1 && keyField != col['name']) {
           fieldValues.remove(keyField);
           keyField = col['name'];
+          _keyFields[table] = keyField;
           fieldValues[keyField] = null;
           fields.first = keyField;
         } else {
@@ -626,7 +647,23 @@ class _DBInterface {
             fieldValues[col['name']] = col['dflt_value'];
           } else {
             if (col['notnull'] == 1) {
-              //TODO Determine the default by datatype.
+              switch (col['type']) {
+                case 'Long':
+                  {
+                    fldValue = 0.0;
+                  }
+                  break;
+                case 'INTEGER':
+                  {
+                    fldValue = 0;
+                  }
+                  break;
+                default:
+                  {
+                    fldValue = '';
+                  }
+              }
+              fieldValues[col['name']] = fldValue;
             } else {
               fieldValues[col['name']] = null;
             }
@@ -636,21 +673,28 @@ class _DBInterface {
 
       _fields[table] = fields;
 
-      _fldValues[table] = fieldValues;
+//      _fldValues[table] = fieldValues;
 
-      _newRec[table] = Map();
+      _newRec[table] = Map<String, dynamic>();
 
       /// Make a copy as an 'empty' record.
-      _newRec[table].addEntries(_fldValues[table].entries);
+      _newRec[table]
+          .addEntries(fieldValues.entries); //_fldValues[table].entries);
     }
   }
 
   Future<String> get localPath async {
-    if(_path == null) {
-      var directory = await getApplicationDocumentsDirectory();
-      _path = directory.path;
+    if (_path == null) {
+      try {
+        Directory directory = await getApplicationDocumentsDirectory();
+        _path = directory.path;
+      } catch (e) {
+        ex = e;
+        _path = '';
+      }
     }
     return _path;
   }
+
   String _path;
 }